- Refactor, split classes - Add ChatGPT
