+++ Image2List/ImageProcessing/VisionFiles/GroceryVocabulary.swift
